refactor: extract browser_launch_env into proxy_env module (line-limit)
Moves the proxy-stripping helper out of ziniao_utils.py (back under the
800-line limit) into app/browser/proxy_env.py — no circular deps (it only
needs os). ziniao_utils re-imports it. Documents the macOS open-a/launchd
limitation (env-strip only bites on the Windows/Linux exec paths).
